pyproject.toml
[tool.poetry]
name = "fastapi-graphql"
version = "0.1.0"
description = ""
authors = ["Anil Khatri <anil.soccer.khatri@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.61.2"
uvicorn = "^0.12.2"
orator = "^0.9.9"
psycopg2-binary = "^2.8.6"
graphene = "^2.1.8"
graphene-pydantic = "0.1.0"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.790"
flake8 = "^3.8.4"
autoflake = "^1.4"
isort = "^5.6.4"
black = "^20.8b1"
pre-commit = "^2.8.2"

[tool.black]
line-length = 99
skip-string-normalization = true
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | \.env
    | _build
    | buck-out
    | build
    | dist
  )/
  | /migrations/                # ignore migrations
  | /site-packages/             # ignore thrid-party packages
)
'''

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
